+++ internal/core/domain/receiver.go
@@ -1,0 +1,8 @@
+package domain
+
+type ReceiverType = string
+
+const (
+	RequestReceiverType ReceiverType = "request"
+	ImapReceiverType    ReceiverType = "imap"
+)